Add Authenticate step to mailing list sign up
Make the `Authenticate` step generic and re-use it in the event and mailing
list sign ups.
