+++ app/helpers/session_helper.rb
@@ -3,4 +3,8 @@ module SessionHelper
     session[:events] ||= {}
     session[:events][params[:event_id]] ||= {}
   end
+
+  def mailing_list_session
+    session[:mailinglist] ||= {}
+  end
 end
