+++ app/views/wizard/steps/_authenticate.html.erb
@@ -1,0 +1,9 @@
+<%
+  hint_text = t("helpers.hint.wizard_steps_authenticate.timed_one_time_password.text", 
+    link: link_to(t("helpers.hint.wizard_steps_authenticate.timed_one_time_password.link"), resend_verification_path))
+  hint_text += "<br /><b>#{t("helpers.hint.wizard_steps_authenticate.timed_one_time_password.resent")}</b>" if params[:verification_resent]
+%>
+
+<%= f.govuk_text_field :timed_one_time_password, 
+label: { text: t('helpers.label.wizard_steps_authenticate.timed_one_time_password', email: email) }, 
+hint_text: hint_text.html_safe %>
