+++ spec/factories/wIzard/authenticate_factory.rb
